rename pdf2json to grobid2json and make room for spp2json
